requirements.txt
# Core deep learning framework
torch>=1.9.0
torchvision>=0.10.0
torchaudio>=0.9.0

# Essential scientific computing
numpy>=1.21.0
scipy>=1.7.0
opencv-python>=4.5.0
Pillow>=8.3.0

# Data manipulation and analysis
pandas>=1.3.0
matplotlib>=3.3.0
seaborn>=0.11.0

# Machine learning utilities
scikit-learn>=1.0.0
scikit-image>=0.18.0

# Training and visualization
tensorboard>=2.7.0
tqdm>=4.62.0
wandb>=0.12.0

# Image and video processing
imageio>=2.9.0
h5py>=3.3.0

# KAN-related dependencies (optional, will fallback to MLP if not available)
# efficient-kan  # Uncomment if available

# Additional utilities
psutil>=5.8.0
timm>=0.5.4
albumentations>=1.1.0

# Development and testing
pytest>=6.2.0
black>=21.0.0
flake8>=3.9.0

# For specific model implementations
efficientnet-pytorch>=0.7.1
pretrainedmodels>=0.7.4

# For advanced optimizers and schedulers
torch-optimizer>=0.3.0

# For mixed precision training (included in newer PyTorch versions)
# nvidia-apex  # Optional, for older PyTorch versions

# For better progress bars and logging
rich>=10.0.0
loguru>=0.5.0

# For configuration management
omegaconf>=2.1.0
hydra-core>=1.1.0

# For model optimization
onnx>=1.10.0
onnxruntime>=1.9.0

# For dataset downloading and processing
gdown>=4.0.0
requests>=2.26.0

# Memory profiling
memory-profiler>=0.58.0

# Version pinning for stability
torchmetrics>=0.6.0
